+++ release-calendar/package.json
@@ -22,6 +22,8 @@
     "server-build": "webpack --config webpack.server.config.js",
     "server-serve": "node -r dotenv/config dist/bundle-server.js",
     "setup": "npm run setup-build && npm run setup-serve",
+    "setup-build": "webpack --config webpack.setup.config.js",
+    "setup-serve": "node -r dotenv/config dist/bundle-setup.js",
     "setup-build": "webpack --config ./webpack.setup.config.js",
     "setup-serve": "node -r dotenv/config ./dist/bundle-setup.js",
     "pretest": "npm run lint",
